requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
qdrant-client
langchain-qdrant
pypdf
langchain-huggingface
python-dotenv
fastapi
uvicorn
mangum
mlflow
langsmith
pydantic
python-multipart
streamlit
requests
ragas
datasets
pandas
